move helpers to email layer
