+++ pkg/email/ctat_emails.go
@@ -1,9 +1,18 @@
 package email
 
 import (
+	"context"
 	_ "embed"
+	"fmt"
+	"slices"
+	"strings"
 
+	"github.com/google/uuid"
+	"github.com/samber/lo"
+
+	"github.com/cms-enterprise/mint-app/pkg/models"
 	"github.com/cms-enterprise/mint-app/pkg/shared/emailtemplates"
+	"github.com/cms-enterprise/mint-app/pkg/shared/oddmail"
 )
 
 // CTATSubmittedTemplateName is the template name for the CTAT submitted email.
@@ -102,4 +111,268 @@ type CTATUpdateBodyContent struct {
 	RequestUrgency            string
 	DateAssistanceNeededBy    string
 	UploadedFiles             string
+}
+
+// BuildCTATSubmittedBodyContent assembles the CTAT submitted email body content.
+func BuildCTATSubmittedBodyContent(
+	ctx context.Context,
+	emailService oddmail.EmailService,
+	ctatRequest *models.CTATRequest,
+	getRelatedMINTModels func(context.Context, uuid.UUID) ([]*models.ModelPlan, error),
+	getSupportingDocuments func(context.Context, uuid.UUID) ([]*models.CTATRequestDocument, error),
+) (CTATSubmittedBodyContent, error) {
+	if emailService == nil || ctatRequest == nil {
+		return CTATSubmittedBodyContent{}, nil
+	}
+
+	requesterAccount, err := ctatRequest.RequesterUserAccount(ctx)
+	if err != nil {
+		return CTATSubmittedBodyContent{}, err
+	}
+
+	relatedModels := []*models.ModelPlan{}
+	if getRelatedMINTModels != nil {
+		relatedModels, err = getRelatedMINTModels(ctx, ctatRequest.ID)
+		if err != nil {
+			return CTATSubmittedBodyContent{}, err
+		}
+	}
+
+	documents := []*models.CTATRequestDocument{}
+	if getSupportingDocuments != nil {
+		documents, err = getSupportingDocuments(ctx, ctatRequest.ID)
+		if err != nil {
+			return CTATSubmittedBodyContent{}, err
+		}
+	}
+
+	bodyContent := CTATSubmittedBodyContent{
+		ClientAddress:          emailService.GetConfig().GetClientAddress(),
+		CTATTicketID:           ctatRequest.ID.String(),
+		TicketNumber:           ctatRequest.HumanReadableID(),
+		CMMIGroup:              ctatRequest.CmmiGroup.Humanize(),
+		RelatedMINTModels:      strings.Join(lo.Map(relatedModels, func(item *models.ModelPlan, _ int) string { return item.ModelName }), ", "),
+		TypeOfHelpNeeded:       strings.Join(lo.Map(ctatRequest.TypeOfHelpNeeded, func(item models.CTATHelpNeededType, _ int) string { return item.Humanize() }), ", "),
+		DescribeHelpNeeded:     ctatRequest.DescribeHelpNeeded,
+		RequestUrgency:         ctatRequest.RequestUrgency.Humanize(),
+		DateAssistanceNeededBy: ctatRequest.DateAssistanceNeededBy.Format("01/02/2006"),
+		UploadedFiles:          strings.Join(lo.Map(documents, func(item *models.CTATRequestDocument, _ int) string { return item.FileName }), ", "),
+	}
+
+	if requesterAccount != nil {
+		bodyContent.RequesterName = requesterAccount.CommonName
+		bodyContent.RequesterEmail = requesterAccount.Email
+	}
+
+	if ctatRequest.ContractName != nil {
+		bodyContent.ContractName = *ctatRequest.ContractName
+	}
+
+	if ctatRequest.CmmiGroup == models.CTATCMMIGroupOptionOther && ctatRequest.CmmiGroupOther != nil {
+		bodyContent.CMMIGroup = fmt.Sprintf("%s (%s)", bodyContent.CMMIGroup, *ctatRequest.CmmiGroupOther)
+	}
+
+	if ctatRequest.CmmiDivision != nil {
+		bodyContent.CMMIDivision = ctatRequest.CmmiDivision.Humanize()
+		if *ctatRequest.CmmiDivision == models.CTATCMMIDivisionOptionOther && ctatRequest.CmmiDivisionOther != nil {
+			bodyContent.CMMIDivision = fmt.Sprintf("%s (%s)", bodyContent.CMMIDivision, *ctatRequest.CmmiDivisionOther)
+		}
+	}
+
+	if ctatRequest.ContractActivityType != nil {
+		bodyContent.ContractActivityType = ctatRequest.ContractActivityType.Humanize()
+		if *ctatRequest.ContractActivityType == models.CTATContractActivityTypeOther && ctatRequest.ContractActivityTypeOther != nil {
+			bodyContent.ContractActivityType = fmt.Sprintf("%s (%s)", bodyContent.ContractActivityType, *ctatRequest.ContractActivityTypeOther)
+		}
+	}
+
+	if ctatRequest.ContractType != nil {
+		bodyContent.ContractType = ctatRequest.ContractType.Humanize()
+		if *ctatRequest.ContractType == models.CTATContractTypeOther && ctatRequest.ContractTypeOther != nil {
+			bodyContent.ContractType = fmt.Sprintf("%s (%s)", bodyContent.ContractType, *ctatRequest.ContractTypeOther)
+		}
+	}
+
+	if slices.Contains(ctatRequest.TypeOfHelpNeeded, models.CTATHelpNeededTypeOther) && ctatRequest.TypeOfHelpNeededOther != nil {
+		helpNeededValues := lo.Map(ctatRequest.TypeOfHelpNeeded, func(item models.CTATHelpNeededType, _ int) string {
+			humanized := item.Humanize()
+			if item == models.CTATHelpNeededTypeOther {
+				return fmt.Sprintf("%s (%s)", humanized, *ctatRequest.TypeOfHelpNeededOther)
+			}
+
+			return humanized
+		})
+		bodyContent.TypeOfHelpNeeded = strings.Join(helpNeededValues, ", ")
+	}
+
+	return bodyContent, nil
+}
+
+// SendCTATSubmittedEmail sends the CTAT submitted email to the requester.
+func SendCTATSubmittedEmail(
+	ctx context.Context,
+	emailService oddmail.EmailService,
+	addressBook AddressBook,
+	ctatRequest *models.CTATRequest,
+	getRelatedMINTModels func(context.Context, uuid.UUID) ([]*models.ModelPlan, error),
+	getSupportingDocuments func(context.Context, uuid.UUID) ([]*models.CTATRequestDocument, error),
+) error {
+	if emailService == nil || ctatRequest == nil {
+		return nil
+	}
+
+	subjectContent := CTATSubmittedSubjectContent{
+		TicketNumber: ctatRequest.HumanReadableID(),
+	}
+
+	bodyContent, err := BuildCTATSubmittedBodyContent(
+		ctx,
+		emailService,
+		ctatRequest,
+		getRelatedMINTModels,
+		getSupportingDocuments,
+	)
+	if err != nil {
+		return err
+	}
+	if bodyContent.RequesterEmail == "" {
+		return nil
+	}
+
+	emailSubject, emailBody, err := CTAT.Submitted.GetContent(subjectContent, bodyContent)
+	if err != nil {
+		return err
+	}
+
+	return emailService.Send(
+		addressBook.DefaultSender,
+		[]string{bodyContent.RequesterEmail},
+		nil,
+		emailSubject,
+		"text/html",
+		emailBody,
+	)
+}
+
+func uuidPointersEqual(first *uuid.UUID, second *uuid.UUID) bool {
+	if first == nil && second == nil {
+		return true
+	}
+
+	if first == nil || second == nil {
+		return false
+	}
+
+	return *first == *second
+}
+
+func trimmedStringPointersEqual(first *string, second *string) bool {
+	firstValue := ""
+	if first != nil {
+		firstValue = strings.TrimSpace(*first)
+	}
+
+	secondValue := ""
+	if second != nil {
+		secondValue = strings.TrimSpace(*second)
+	}
+
+	return firstValue == secondValue
+}
+
+// SendCTATUpdateEmail sends the CTAT update email to the requester.
+func SendCTATUpdateEmail(
+	ctx context.Context,
+	emailService oddmail.EmailService,
+	addressBook AddressBook,
+	originalRequest *models.CTATRequest,
+	updatedRequest *models.CTATRequest,
+	getRelatedMINTModels func(context.Context, uuid.UUID) ([]*models.ModelPlan, error),
+	getSupportingDocuments func(context.Context, uuid.UUID) ([]*models.CTATRequestDocument, error),
+) error {
+	if emailService == nil || originalRequest == nil || updatedRequest == nil {
+		return nil
+	}
+
+	statusUpdated := originalRequest.Status != updatedRequest.Status
+	assignedAdminUpdated := !uuidPointersEqual(originalRequest.AssignedAdmin, updatedRequest.AssignedAdmin)
+	progressNotesUpdated := !trimmedStringPointersEqual(originalRequest.Notes, updatedRequest.Notes)
+	resolutionUpdated := !trimmedStringPointersEqual(originalRequest.Resolution, updatedRequest.Resolution)
+	if !statusUpdated && !assignedAdminUpdated && !progressNotesUpdated && !resolutionUpdated {
+		return nil
+	}
+
+	bodySummary, err := BuildCTATSubmittedBodyContent(
+		ctx,
+		emailService,
+		updatedRequest,
+		getRelatedMINTModels,
+		getSupportingDocuments,
+	)
+	if err != nil {
+		return err
+	}
+	if bodySummary.RequesterEmail == "" {
+		return nil
+	}
+
+	assignedAdminAccount, err := updatedRequest.AssignedAdminUserAccount(ctx)
+	if err != nil {
+		return err
+	}
+
+	subjectContent := CTATUpdateSubjectContent{
+		TicketNumber: updatedRequest.HumanReadableID(),
+	}
+
+	bodyContent := CTATUpdateBodyContent{
+		Status:                    updatedRequest.Status.Humanize(),
+		StatusUpdated:             statusUpdated,
+		AssignedTeamMemberUpdated: assignedAdminUpdated,
+		ProgressNotesUpdated:      progressNotesUpdated,
+		ResolutionUpdated:         resolutionUpdated,
+		ClientAddress:             bodySummary.ClientAddress,
+		CTATTicketID:              bodySummary.CTATTicketID,
+		TicketNumber:              bodySummary.TicketNumber,
+		RequesterName:             bodySummary.RequesterName,
+		RequesterEmail:            bodySummary.RequesterEmail,
+		CMMIGroup:                 bodySummary.CMMIGroup,
+		CMMIDivision:              bodySummary.CMMIDivision,
+		RelatedMINTModels:         bodySummary.RelatedMINTModels,
+		ContractActivityType:      bodySummary.ContractActivityType,
+		ContractName:              bodySummary.ContractName,
+		ContractType:              bodySummary.ContractType,
+		TypeOfHelpNeeded:          bodySummary.TypeOfHelpNeeded,
+		DescribeHelpNeeded:        bodySummary.DescribeHelpNeeded,
+		RequestUrgency:            bodySummary.RequestUrgency,
+		DateAssistanceNeededBy:    bodySummary.DateAssistanceNeededBy,
+		UploadedFiles:             bodySummary.UploadedFiles,
+	}
+
+	if assignedAdminAccount != nil {
+		bodyContent.AssignedTeamMemberName = assignedAdminAccount.CommonName
+		bodyContent.AssignedTeamMemberEmail = assignedAdminAccount.Email
+	}
+
+	if updatedRequest.Notes != nil {
+		bodyContent.ProgressNotes = *updatedRequest.Notes
+	}
+
+	if updatedRequest.Resolution != nil {
+		bodyContent.Resolution = *updatedRequest.Resolution
+	}
+
+	emailSubject, emailBody, err := CTAT.Update.GetContent(subjectContent, bodyContent)
+	if err != nil {
+		return err
+	}
+
+	return emailService.Send(
+		addressBook.DefaultSender,
+		[]string{bodySummary.RequesterEmail},
+		nil,
+		emailSubject,
+		"text/html",
+		emailBody,
+	)
 }
