+++ pkg/graph/resolvers/ctat_request.go
@@ -3,8 +3,6 @@ package resolvers
 import (
 	"context"
 	"fmt"
-	"slices"
-	"strings"
 
 	"github.com/google/uuid"
 	"github.com/jmoiron/sqlx"
@@ -111,7 +109,15 @@ func CTATRequestAdminUpdate(
 
 	if emailService != nil {
 		go func() {
-			sendEmailErr := sendCTATUpdateEmail(ctx, emailService, addressBook, &originalRequest, updatedRequest)
+			sendEmailErr := email.SendCTATUpdateEmail(
+				ctx,
+				emailService,
+				addressBook,
+				&originalRequest,
+				updatedRequest,
+				CTATRelatedMINTModelsGetByCTATRequestIDLOADER,
+				CTATRequestDocumentGetByCTATRequestIDLOADER,
+			)
 			if sendEmailErr != nil {
 				logger.Error(
 					"failed to send CTAT update email",
@@ -215,7 +221,14 @@ func CTATRequestCreate(
 
 	if emailService != nil {
 		go func() {
-			sendEmailErr := sendCTATSubmittedEmail(ctx, emailService, addressBook, createdRequest)
+			sendEmailErr := email.SendCTATSubmittedEmail(
+				ctx,
+				emailService,
+				addressBook,
+				createdRequest,
+				CTATRelatedMINTModelsGetByCTATRequestIDLOADER,
+				CTATRequestDocumentGetByCTATRequestIDLOADER,
+			)
 			if sendEmailErr != nil {
 				logger.Error(
 					"failed to send CTAT submitted email",
@@ -227,245 +240,6 @@ func CTATRequestCreate(
 	}
 
 	return createdRequest, nil
-}
-
-func buildCTATSubmittedBodyContent(
-	ctx context.Context,
-	emailService oddmail.EmailService,
-	ctatRequest *models.CTATRequest,
-) (email.CTATSubmittedBodyContent, error) {
-	if emailService == nil || ctatRequest == nil {
-		return email.CTATSubmittedBodyContent{}, nil
-	}
-
-	requesterAccount, err := ctatRequest.RequesterUserAccount(ctx)
-	if err != nil {
-		return email.CTATSubmittedBodyContent{}, err
-	}
-
-	relatedModels, err := CTATRelatedMINTModelsGetByCTATRequestIDLOADER(ctx, ctatRequest.ID)
-	if err != nil {
-		return email.CTATSubmittedBodyContent{}, err
-	}
-
-	documents, err := CTATRequestDocumentGetByCTATRequestIDLOADER(ctx, ctatRequest.ID)
-	if err != nil {
-		return email.CTATSubmittedBodyContent{}, err
-	}
-
-	bodyContent := email.CTATSubmittedBodyContent{
-		ClientAddress:          emailService.GetConfig().GetClientAddress(),
-		CTATTicketID:           ctatRequest.ID.String(),
-		TicketNumber:           ctatRequest.HumanReadableID(),
-		CMMIGroup:              ctatRequest.CmmiGroup.Humanize(),
-		RelatedMINTModels:      strings.Join(lo.Map(relatedModels, func(item *models.ModelPlan, _ int) string { return item.ModelName }), ", "),
-		TypeOfHelpNeeded:       strings.Join(lo.Map(ctatRequest.TypeOfHelpNeeded, func(item models.CTATHelpNeededType, _ int) string { return item.Humanize() }), ", "),
-		DescribeHelpNeeded:     ctatRequest.DescribeHelpNeeded,
-		RequestUrgency:         ctatRequest.RequestUrgency.Humanize(),
-		DateAssistanceNeededBy: ctatRequest.DateAssistanceNeededBy.Format("01/02/2006"),
-		UploadedFiles:          strings.Join(lo.Map(documents, func(item *models.CTATRequestDocument, _ int) string { return item.FileName }), ", "),
-	}
-
-	if requesterAccount != nil {
-		bodyContent.RequesterName = requesterAccount.CommonName
-		bodyContent.RequesterEmail = requesterAccount.Email
-	}
-
-	if ctatRequest.ContractName != nil {
-		bodyContent.ContractName = *ctatRequest.ContractName
-	}
-
-	if ctatRequest.CmmiGroup == models.CTATCMMIGroupOptionOther && ctatRequest.CmmiGroupOther != nil {
-		bodyContent.CMMIGroup = fmt.Sprintf("%s (%s)", bodyContent.CMMIGroup, *ctatRequest.CmmiGroupOther)
-	}
-
-	if ctatRequest.CmmiDivision != nil {
-		bodyContent.CMMIDivision = ctatRequest.CmmiDivision.Humanize()
-		if *ctatRequest.CmmiDivision == models.CTATCMMIDivisionOptionOther && ctatRequest.CmmiDivisionOther != nil {
-			bodyContent.CMMIDivision = fmt.Sprintf("%s (%s)", bodyContent.CMMIDivision, *ctatRequest.CmmiDivisionOther)
-		}
-	}
-
-	if ctatRequest.ContractActivityType != nil {
-		bodyContent.ContractActivityType = ctatRequest.ContractActivityType.Humanize()
-		if *ctatRequest.ContractActivityType == models.CTATContractActivityTypeOther && ctatRequest.ContractActivityTypeOther != nil {
-			bodyContent.ContractActivityType = fmt.Sprintf("%s (%s)", bodyContent.ContractActivityType, *ctatRequest.ContractActivityTypeOther)
-		}
-	}
-
-	if ctatRequest.ContractType != nil {
-		bodyContent.ContractType = ctatRequest.ContractType.Humanize()
-		if *ctatRequest.ContractType == models.CTATContractTypeOther && ctatRequest.ContractTypeOther != nil {
-			bodyContent.ContractType = fmt.Sprintf("%s (%s)", bodyContent.ContractType, *ctatRequest.ContractTypeOther)
-		}
-	}
-
-	if slices.Contains(ctatRequest.TypeOfHelpNeeded, models.CTATHelpNeededTypeOther) && ctatRequest.TypeOfHelpNeededOther != nil {
-		helpNeededValues := lo.Map(ctatRequest.TypeOfHelpNeeded, func(item models.CTATHelpNeededType, _ int) string {
-			humanized := item.Humanize()
-			if item == models.CTATHelpNeededTypeOther {
-				return fmt.Sprintf("%s (%s)", humanized, *ctatRequest.TypeOfHelpNeededOther)
-			}
-
-			return humanized
-		})
-		bodyContent.TypeOfHelpNeeded = strings.Join(helpNeededValues, ", ")
-	}
-
-	return bodyContent, nil
-}
-
-func sendCTATSubmittedEmail(
-	ctx context.Context,
-	emailService oddmail.EmailService,
-	addressBook email.AddressBook,
-	ctatRequest *models.CTATRequest,
-) error {
-	if emailService == nil || ctatRequest == nil {
-		return nil
-	}
-
-	subjectContent := email.CTATSubmittedSubjectContent{
-		TicketNumber: ctatRequest.HumanReadableID(),
-	}
-
-	bodyContent, err := buildCTATSubmittedBodyContent(ctx, emailService, ctatRequest)
-	if err != nil {
-		return err
-	}
-	if bodyContent.RequesterEmail == "" {
-		return nil
-	}
-
-	emailSubject, emailBody, err := email.CTAT.Submitted.GetContent(subjectContent, bodyContent)
-	if err != nil {
-		return err
-	}
-
-	return emailService.Send(
-		addressBook.DefaultSender,
-		[]string{bodyContent.RequesterEmail},
-		nil,
-		emailSubject,
-		"text/html",
-		emailBody,
-	)
-}
-
-func uuidPointersEqual(first *uuid.UUID, second *uuid.UUID) bool {
-	if first == nil && second == nil {
-		return true
-	}
-
-	if first == nil || second == nil {
-		return false
-	}
-
-	return *first == *second
-}
-
-func trimmedStringPointersEqual(first *string, second *string) bool {
-	firstValue := ""
-	if first != nil {
-		firstValue = strings.TrimSpace(*first)
-	}
-
-	secondValue := ""
-	if second != nil {
-		secondValue = strings.TrimSpace(*second)
-	}
-
-	return firstValue == secondValue
-}
-
-func sendCTATUpdateEmail(
-	ctx context.Context,
-	emailService oddmail.EmailService,
-	addressBook email.AddressBook,
-	originalRequest *models.CTATRequest,
-	updatedRequest *models.CTATRequest,
-) error {
-	if emailService == nil || originalRequest == nil || updatedRequest == nil {
-		return nil
-	}
-
-	// we only want to send the email if there was an update, so compute those ahead of time
-	// note: we won't send an email for a whitespace-only change, but we will still save the change in the DB
-	statusUpdated := originalRequest.Status != updatedRequest.Status
-	assignedAdminUpdated := !uuidPointersEqual(originalRequest.AssignedAdmin, updatedRequest.AssignedAdmin)
-	progressNotesUpdated := !trimmedStringPointersEqual(originalRequest.Notes, updatedRequest.Notes)
-	resolutionUpdated := !trimmedStringPointersEqual(originalRequest.Resolution, updatedRequest.Resolution)
-	if !statusUpdated && !assignedAdminUpdated && !progressNotesUpdated && !resolutionUpdated {
-		return nil
-	}
-
-	bodySummary, err := buildCTATSubmittedBodyContent(ctx, emailService, updatedRequest)
-	if err != nil {
-		return err
-	}
-	if bodySummary.RequesterEmail == "" {
-		return nil
-	}
-
-	assignedAdminAccount, err := updatedRequest.AssignedAdminUserAccount(ctx)
-	if err != nil {
-		return err
-	}
-
-	subjectContent := email.CTATUpdateSubjectContent{
-		TicketNumber: updatedRequest.HumanReadableID(),
-	}
-
-	bodyContent := email.CTATUpdateBodyContent{
-		Status:                    updatedRequest.Status.Humanize(),
-		StatusUpdated:             statusUpdated,
-		AssignedTeamMemberUpdated: assignedAdminUpdated,
-		ProgressNotesUpdated:      progressNotesUpdated,
-		ResolutionUpdated:         resolutionUpdated,
-		ClientAddress:             bodySummary.ClientAddress,
-		CTATTicketID:              bodySummary.CTATTicketID,
-		TicketNumber:              bodySummary.TicketNumber,
-		RequesterName:             bodySummary.RequesterName,
-		RequesterEmail:            bodySummary.RequesterEmail,
-		CMMIGroup:                 bodySummary.CMMIGroup,
-		CMMIDivision:              bodySummary.CMMIDivision,
-		RelatedMINTModels:         bodySummary.RelatedMINTModels,
-		ContractActivityType:      bodySummary.ContractActivityType,
-		ContractName:              bodySummary.ContractName,
-		ContractType:              bodySummary.ContractType,
-		TypeOfHelpNeeded:          bodySummary.TypeOfHelpNeeded,
-		DescribeHelpNeeded:        bodySummary.DescribeHelpNeeded,
-		RequestUrgency:            bodySummary.RequestUrgency,
-		DateAssistanceNeededBy:    bodySummary.DateAssistanceNeededBy,
-		UploadedFiles:             bodySummary.UploadedFiles,
-	}
-
-	if assignedAdminAccount != nil {
-		bodyContent.AssignedTeamMemberName = assignedAdminAccount.CommonName
-		bodyContent.AssignedTeamMemberEmail = assignedAdminAccount.Email
-	}
-
-	if updatedRequest.Notes != nil {
-		bodyContent.ProgressNotes = *updatedRequest.Notes
-	}
-
-	if updatedRequest.Resolution != nil {
-		bodyContent.Resolution = *updatedRequest.Resolution
-	}
-
-	emailSubject, emailBody, err := email.CTAT.Update.GetContent(subjectContent, bodyContent)
-	if err != nil {
-		return err
-	}
-
-	return emailService.Send(
-		addressBook.DefaultSender,
-		[]string{bodySummary.RequesterEmail},
-		nil,
-		emailSubject,
-		"text/html",
-		emailBody,
-	)
 }
 
 type ctatRequestDocumentUpload struct {
